requirements.txt
customtkinter>=5.2
pywin32>=306
Pillow>=10.0
